finish with global navBar
